requirements.txt
cudatoolkit==10.2 #
numpy==1.18.1 # 오류 -> 최신버전
opencv==3.4.2
pandas==1.2.0 #
pillow==7.0 #
pip==20.0 # 오류 -> 최신버전
python==3.7.6
pytorch==1.6.0 
torchvision==0.7 
tensorboard==2.1 #
matplotlib==3.1.3 # 
scikit-learn==0.22.1 #
pyyaml==5.3.1 #
nvidia-apex==0.1
transformers==4.1.1 # 
